Styling Changes
Changed styling based on the Figma. for the index (home) page. Added some personal color changed as well. added a grey to match the logo, and used that same background for the boxes.

Looking to try and fit red into the color scheme somewhere, if you are working on the homepage and see this please play around with the colors and see what you can do.